Lines are now broken into different types of words.
Before, as a quick hack, all lines were send to the backend as a single
plain word. We now parse through each line, recognize the form of words,
and send then individually to the backend.

There are some temporary hacks for some problems that aren't solved yet.
Also, there are a lot of special cases, so the code is ugly and hard to
be sure it's correct. I think this might be unavoidable, but it's worth
investigating.